Chore: 코드 분리 및 드랍다운 사용 수정
